mevbot/strategies/jit_liquidity.py
"""Just-In-Time (JIT) Liquidity strategy implementation"""
from typing import Dict, List, Optional, Tuple
from web3 import Web3
from eth_typing import Address
import asyncio
import logging
from dataclasses import dataclass
import numpy as np
from ..core.simd import SIMDProcessor
from ..core.flashbots import FlashbotsManager
from ..core.memory import MemoryManager
from ..core.flash_loan import FlashLoanManager, FlashLoanParams, AAVE_V3

logger = logging.getLogger(__name__)

@dataclass
class JITParams:
    pool_address: Address
    token_address: Address
    target_tx_hash: str
    liquidity_amount: int
    expected_profit: float
    add_liquidity_tx: Optional[Dict] = None
    remove_liquidity_tx: Optional[Dict] = None

class JITLiquidityStrategy:
    def __init__(self, w3: Web3, flashbots: FlashbotsManager, memory: MemoryManager, flash_loan: FlashLoanManager, config: Dict):
        self.w3 = w3
        self.flashbots = flashbots
        self.memory = memory
        self.flash_loan = flash_loan
        self.config = config
        self.simd = SIMDProcessor()
        
        # Performance tracking
        self.total_attempts = 0
        self.successful_attempts = 0
        self.total_profit = 0.0
        self.total_gas_used = 0
        
        # Initialize memory regions
        self.pool_cache = self.memory.allocate_region('jit_pool_cache', 256)  # 256MB cache
        self.tx_cache = self.memory.allocate_region('jit_tx_cache', 128)     # 128MB cache
        
    async def analyze_opportunity(
        self,
        target_tx: Dict,
        pool_info: Dict
    ) -> Optional[JITParams]:
        """Analyze if a transaction is suitable for JIT liquidity"""
        try:
            # Extract parameters
            pool = pool_info['address']
            token = pool_info['token']
            target_amount = int(target_tx.get('value', 0))
            
            # Quick checks
            if not self._is_profitable_target(target_tx, pool_info):
                return None
                
            # Calculate optimal liquidity amount using SIMD
            liquidity_amount = await self._calculate_optimal_liquidity(
                pool,
                token,
                target_amount
            )
            
            if not liquidity_amount:
                return None
                
            # Simulate JIT transactions
            success, profit = await self._simulate_jit(
                pool,
                token,
                liquidity_amount,
                target_tx
            )
            
            if not success or profit <= 0:
                return None
                
            return JITParams(
                pool_address=pool,
                token_address=token,
                target_tx_hash=target_tx['hash'],
                liquidity_amount=liquidity_amount,
                expected_profit=profit
            )
            
        except Exception as e:
            logger.error(f"Error analyzing JIT opportunity: {str(e)}")
            return None
            
    async def execute_opportunity(self, params: JITParams) -> bool:
        """Execute a JIT liquidity opportunity using flash loans"""
        try:
            # Get optimal flash loan provider
            provider = await self.flash_loan.get_optimal_provider(
                params.token_address,
                params.liquidity_amount
            )
            
            if not provider:
                logger.error("No suitable flash loan provider found")
                return False
                
            # Build flash loan parameters
            flash_params = FlashLoanParams(
                provider=provider,
                token_address=params.token_address,
                amount=params.liquidity_amount,
                callback_data=self._build_callback_data(params),
                expected_profit=params.expected_profit
            )
            
            # Execute flash loan bundle
            success, receipt = await self.flash_loan.execute_flash_loan(flash_params)
            
            if success:
                self.successful_attempts += 1
                self.total_profit += params.expected_profit
                
            self.total_attempts += 1
            return success
            
        except Exception as e:
            logger.error(f"Error executing JIT opportunity: {str(e)}")
            return False
            
    def _is_profitable_target(self, tx: Dict, pool_info: Dict) -> bool:
        """Quick check if a transaction is worth JIT liquidity"""
        try:
            # Check transaction value
            min_value = self.config.get('min_target_value', 1e18)  # 1 ETH default
            if int(tx.get('value', 0)) < min_value:
                return False
                
            # Check pool TVL
            min_tvl = self.config.get('min_pool_tvl', 100e18)  # 100 ETH default
            if pool_info.get('tvl', 0) < min_tvl:
                return False
                
            # Check gas price
            max_gas_price = self.config.get('max_gas_price', 100e9)  # 100 GWEI default
            if int(tx.get('gasPrice', 0)) > max_gas_price:
                return False
                
            return True
            
        except Exception as e:
            logger.error(f"Error checking profitability: {str(e)}")
            return False
            
    async def _calculate_optimal_liquidity(
        self,
        pool: Address,
        token: Address,
        target_amount: int
    ) -> Optional[int]:
        """Calculate optimal liquidity amount using SIMD"""
        try:
            # Get current pool state
            reserves = await self._get_reserves(pool)
            if not reserves:
                return None
                
            # Use SIMD for parallel price impact calculation
            test_amounts = np.array([
                target_amount * 0.5,  # 50%
                target_amount * 1.0,  # 100%
                target_amount * 2.0,  # 200%
                target_amount * 5.0   # 500%
            ])
            
            impacts = self.simd.calculate_price_impacts(
                reserves,
                test_amounts
            )
            
            # Find amount with best profit potential
            best_amount = None
            min_impact = float('inf')
            
            for amount, impact in zip(test_amounts, impacts):
                if impact < min_impact:
                    min_impact = impact
                    best_amount = int(amount)
                    
            return best_amount
            
        except Exception as e:
            logger.error(f"Error calculating optimal liquidity: {str(e)}")
            return None
            
    async def _simulate_jit(
        self,
        pool: Address,
        token: Address,
        liquidity_amount: int,
        target_tx: Dict
    ) -> Tuple[bool, float]:
        """Simulate JIT transactions to calculate profit"""
        try:
            # Build simulation transactions
            params = JITParams(
                pool_address=pool,
                token_address=token,
                target_tx_hash=target_tx['hash'],
                liquidity_amount=liquidity_amount,
                expected_profit=0
            )
            
            add_tx, remove_tx = await self._build_jit_txs(params)
            if not add_tx or not remove_tx:
                return False, 0
                
            # Simulate bundle
            success, profit = await self.flashbots.simulate_bundle([
                add_tx,
                {'hash': target_tx['hash']},
                remove_tx
            ])
            
            return success, profit
            
        except Exception as e:
            logger.error(f"Error simulating JIT: {str(e)}")
            return False, 0
            
    async def _build_jit_txs(self, params: JITParams) -> Tuple[Optional[Dict], Optional[Dict]]:
        """Build add and remove liquidity transactions"""
        try:
            # Get current nonce
            nonce = await self.w3.eth.get_transaction_count(
                self.config['address']
            )
            
            # Build add liquidity transaction
            add_tx = {
                'from': self.config['address'],
                'to': params.pool_address,
                'value': params.liquidity_amount,
                'nonce': nonce,
                'gas': 300000,  # Estimate
                'maxFeePerGas': await self.flashbots.get_max_fee(),
                'maxPriorityFeePerGas': await self.flashbots.get_priority_fee()
            }
            
            # Build remove liquidity transaction
            remove_tx = {
                'from': self.config['address'],
                'to': params.pool_address,
                'value': 0,
                'nonce': nonce + 1,
                'gas': 300000,  # Estimate
                'maxFeePerGas': await self.flashbots.get_max_fee(),
                'maxPriorityFeePerGas': await self.flashbots.get_priority_fee()
            }
            
            return add_tx, remove_tx
            
        except Exception as e:
            logger.error(f"Error building JIT transactions: {str(e)}")
            return None, None
            
    async def _get_reserves(self, pool: Address) -> Optional[Tuple[int, int]]:
        """Get pool reserves"""
        try:
            # Check cache first
            cached = self._get_from_cache(str(pool))
            if cached:
                return cached
                
            # Get from chain
            contract = self.w3.eth.contract(
                address=pool,
                abi=self.config['pool_abi']
            )
            
            reserves = await contract.functions.getReserves().call()
            
            # Cache result
            self._cache_reserves(str(pool), reserves)
            
            return reserves
            
        except Exception as e:
            logger.error(f"Error getting reserves: {str(e)}")
            return None
            
    def _get_from_cache(self, key: str) -> Optional[Tuple[int, int]]:
        """Get data from memory-mapped cache"""
        try:
            if not self.pool_cache:
                return None
                
            # Implementation depends on cache structure
            return None  # TODO: Implement
            
        except Exception as e:
            logger.error(f"Error reading from cache: {str(e)}")
            return None
            
    def _cache_reserves(self, key: str, reserves: Tuple[int, int]):
        """Cache reserves in memory-mapped region"""
        try:
            if not self.pool_cache:
                return
                
            # Implementation depends on cache structure
            pass  # TODO: Implement
            
        except Exception as e:
            logger.error(f"Error writing to cache: {str(e)}")

    def _build_callback_data(self, params: JITParams) -> bytes:
        """Build callback data for flash loan"""
        try:
            # Implementation depends on flash loan protocol
            pass  # TODO: Implement
            
        except Exception as e:
            logger.error(f"Error building callback data: {str(e)}")
            return b''
